better handling of canceling tasks on handicapped Windows

+++ freppledb/execute/views.py
@@ -432,7 +432,15 @@ def CancelTask(request, taskid):
     task.save(using=request.database)
     return HttpResponse(content="OK")
   except PermissionError:
-    return HttpResponseServerError("No permission to kill this task")
+    if os.name == 'nt':
+      # Windows doesn't report us why it failed. We just clean things up.
+      task.message = 'Canceled process'
+      task.processid = None
+      task.status = 'Canceled'
+      task.save(using=request.database)
+      return HttpResponse(content="OK")
+    else:
+      return HttpResponseServerError("No permission to kill this task")
   except Exception as e:
     if os.name == 'nt':
       # Windows doesn't report us why it failed. We just clean things up.
